Refactor plugin destroy() and destroy_ui() methods
Instead of calling the destroy() method directly and have it call
destroy_ui() in turn, add a destroy_plugin() method that calls
destroy() followed by the code from the destroy_ui() method. This
way the destroy() method of a plugin can be empty if there is nothing
to destroy. This matches better how the start() and stop() methods
work.

Also move the try/catch/pass block from the callsides of the destroy()
method inside the new destroy_plugin() method.