red-brick: Add a lock to the script manager upload function
Otherwise it might happen that a second execute_script call tries to
upload the script while the first execute_script call is in the process
of uploading it. If the first call then tries to execute the script
while the seconds has it open for writing then exec() will fail with
E_CANNOT_EXEC because a file open for writing cannot be executed.

Now script upload is protected by a lock, so two calls that would have
uploaded the script at the same time will now be serialized. Which fixes
the problem.