red-brick: Try to read script results incremental
Incremental read will allow to workaround the problems with the fixed
pipe size and will allow to show the output of potentially longer
running scripts while it is produced instead of waiting until the script
finished.

The API for this is in place in the redapid, but the usage is more
complex because there are now three callbacks to deal with during
the script's runtime.

Because this is complex delay it until after the first RED Brick
release instead of risking to break things now so near to the planned
release date.