Completely replace the usage of the button pressed signal with the clicked signal
The pressed signal is emitted in mouse down. The clicked signal is emitted
on mouse up. The normal behavior of a button it to trigger on mouse up.
That's why the clicked signal should be prefered over the pressed signal.

Also the pressed signal seems to be the cause for the problems with the
connect button.