red-brick: Fix progess logic in ChunkedDownloaderBase
The set_progress_value function might call QProgressDialog.setValue
which in turn calls QApplication.processEvents which then can result
in delivering queued signals.

Because of this the last_download_size assignment has to be done
before the set_progress_value call otherwise there might be other
status callbacks devlivered which then results in wrong progress
calculations because the assumptions about the last_download_size
are not vaild anymore.

Also keep track of the current progress value locally instead of
having an abstract getter for it.