red-brick: Rework serviceprovider database handling
Only run the build script on Linux, because the ISO 3166 database
is Linux specific. This means that serviceprovider_data.py cannot be
generated easily on Windows or Mac OS X. Therefore, add the file to
git.

Replace xmltodict and pycountry with ElementTree to avoid new external
dependencies.

Also use pformat to output the serviceprovider_data.py files in a
multi-line format which is better for git diffs.