red-brick: Make filesystem resize more robust
Also look at the actual filesystem size to check if it can be expanded.

Move the actual resize work from rc.local to a dedicated init.d script
that runs before brickd and redapid to ensure that the RED Brick doesn't
show up in brickv until the resize is done.