red: Fix config parser usage
The code expected the get function to take section, option and fallback
as parameters. But this is not the official API. Change the get function
to make it work as expected.

Also try to import ConfigParser first, as the configparser spelling can
be found on Python 2 as well and will give trouble.