Remove unnecessary device-factory
The device-factory can get out-of-sync between brickv.git and
generators.git if devices are locally marked as released for
testing purposes while generating the bindings for brickv.git.
This results in devices being marked as released by accident
in brickv.git that are actually not even commited to brickv.git.