samba: Also set security flag along with boot-from-flash flag
This allows Bricks to properly boot from flash, after entering
bootloader-mode by holding the erase button during a power-cycle.

Before, the Brick had to be power-cycled once after flashing to
boot correctly. But this has the downside that entering
bootloader-mode by holding the erase button during a power-cycle
now results in the Brick being in an zombie state between bootloader
and firmware for 15 seconds before it properly enters bootloader-mode.