Introduce generic connections concept for Infos
Every DeviceInfo can now have >= 0 conncetions and up to 1 reverse_connection.
These connections are used to describe the topology of the system.

With the introduction of Bricklets directly connected to a Brick Daemon we
need this more generic approach, since the old strict approach (Bricks have
exactly 2 or 4 Bricklets etc) does notwork anymore. With the new approach
we were also able to remove the Isolator Bricklet hacks, since it integrates
itself automatically (reverse_connection to Brick and connection to isolated
Bricklet).