Reworked handling of uncaught exceptions.
The brick viewer can now be started with the --error-report flag.
In this mode it will read an error report from stdin, display the
report akin to a QErrorMessage and return 1 if errors of this type
should be shown again or 0 if not.

The "normal" brick viewer will put any uncaught exception (f.e. on
missing timeout handling in plugins) in a queue, which is read by
another thread. This thread will run another brickv instance with
the --error-report flag to show the error to the user.

This approach allows the main (GUI) thread to continue on errors, but
will show errors one by one to the user.