Don't delay async-calls using a timer
This can make the async-call jump the session boundary.

In this particular case the WIFI Extension 2.0 version getter call
was delayed by 2 seconds using a QTimer. If disconnect was clicked
during this delay, then a new async-call session would start, but the
QTimer was still running triggering the async-call for the old Master
Brick plugin in the new async-call session. This async-call then
called the WIFI Extension 2.0 version getter that in turn failed with
a not-connected error. This made the async-call thread drop the rest
of the queue. If the connect button was clicked at the right time
between the addition of the WIFI Extension 2.0 version async-call to
the queue and the failure of the WIFI Extension 2.0 version async-call
then the connect async-call was dropped silently from the queue leaving
the GUI in the connecting state without a way for the user to recover
from this.

Fix this by moving the delay logic into the async-call logic to record
the correct async-call session for the WIFI Extension 2.0 version getter.
This results in correctly ignoring delayed async-calls from old sessions.