Improve pyinstaller distutils issue work-around.
Instead of distributing distutils.version, really fix the issue:
Before virtualenv 16.4.0 distutils_path was set to the directory distutils is found in.
Since virtualenv 16.4.0 distutils_path points to the __init__.py file inside this directory.
PyInstaller's distutils-hook still expectes the old format; fake it here.
Also under macOS the distutils_path attribute is not available, but the work-around seems
to be not necessary here.