flashing: Don't try to guess which serial port is a flashable brick.
If the user opens the flashing window by clicking a brick's update
button, it is not in bootloader mode. The old behaviour would then
always select none or the wrong serial port if another (e.g. a RED
Brick's serial console) is available. The pre-selection is removed
completely as other behaviour would imply, that we are 100 percent
sure which port is the correct one.