plot-widget: Resolve sampling race condition with callback emulator
The callback emulator samples the value into a variable and the plot
widget samples the value from this variable. Both sampling processes
normally run at 10 Hz but are not synchronized. This can result in
sample loss and sample duplication.

Avoid this in the plot widget by keeping track of the original sample
timing created by the callback emulator, while also compensating for
jitter in the x-axis rendering and motion.