mainwindow: Don't reset the RED Brick plugin on changed fw version.
The RED Brick reports the redapid version in enumerations. However the
plugin sets the firmware_version_installed in the device_info to the
image version. Enumerations triggered later (for example by the server
room monitoring logic) will then re-report the redapid version. This
should not trigger the plugin reset.

If the firmware version really changes, the RED Brick has to reboot
anyway.