Avoid potential too-many-arguments error
The QAction triggered signal has a toggled argument, but the slot
connected to it takes no arguments. Normally this should be okay,
but there seems to be cases where it does not. Here it might the
functools.partial that creates the problem.

Avoid this situation by just blindly consuming extra arguments.