data-logger: Fix slow memory leak in data tab
It seems that adding rows to a QStandardItemModel attached to a
QTableView slowly leaks memory. Depending on the data rate this
leak takes hours to days to become critical. The exact cause is
unknown, but it seems to be in PyQt5 or Qt5 itself.

Replacing the QStandardItemModel with a custom model fixes the leak.