Improve handling of "wrapped C/C++ object has been deleted" exception
The CallbackEmulator instance was not a child of its parent plugin.
This allowed it to outlive its parent plugin. Set up the parent child
relationship to ensure that the wrapped C/C++ objects are deleted
in sync.

Before, the CallbackEmulator could emit the result signal to a plugin
while the plugin's C/C++ object was already deleted. The exception would
then be raised in the slot while trying to access the already deleted
object.

With the parent child relationship in place this exception should now
happen while emitting the signal, allowing to handle it centralized it
the CallbackEmulator itself.