error-reporter: Make compatible with python 3.10.
traceback.format_exception was modified to take an exception, but still
supports value and tb for backwards compatibility. However etype was
renamed to exc and changed to a positional only parameter. Luckily, the
old signature had exc (then etype) as the first argument, so this version
works with older python versions as well as 3.10.