imu-v2/3: Decouple plot timestamping from signal/slot execution delay
Let the callback emulator directly feed the curve value wrapper without
a signal/slot in the call chain.

It seems that on a Raspberry Pi there can be under certain conditions
be up to 1 sec delay in the execution of a slot. The acquisition and
the timestamping of the plot data is separated by a callback emulator
signal/slot connection. This results in wrong timestamps due to the
strange slot execution delay. Move the timestamping before the slot
execution to workaround the wrong timestamps.