qhexedit: Assume all divisions to be integer divisions
Python 3.10 changed something about implicit float to int conversion
in combination with PyQt5. Before Python 3.10 a float would automatically
be converted to an int if a PyQt5 function would expect an in parameter.
This doesn't happen anymore with Python 3.10, leaving a minefield of
implicit float to int conversions that will now raise TypeErrors.